Socialbar and Navbar Rework
This update was mainly to get the navbar and socialbar working correctly with the new revamp. In doing that I needed to adjust the styling of the main pages.

I added new new dependencies for the updated material ui.

There was some additional styling added to changes to accommodate some of the changes.

Still need to redo project page to make it responsive friendly. Also, possibly looking at moving social, nav, and logo when the width is very small!